requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
dj-rest-auth==6.0.0
django-allauth==0.62.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
channels==4.1.0
daphne==4.1.2
requests==2.31.0